split up the implementation, so it will be easier to deprecate the scan and reuse other parts
