+++ src/storage/compression/dictionary/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library_unity(duckdb_storage_compression_dictionary OBJECT common.cpp
+                  analyze.cpp compression.cpp decompression.cpp)
+set(ALL_OBJECT_FILES
+    ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_storage_compression_dictionary>
+    PARENT_SCOPE)
