+++ src/storage/compression/dictionary/decompression.cpp
@@ -1,0 +1,3 @@
+#include "duckdb/storage/compression/dictionary/decompression.hpp"
+
+namespace duckdb {} // namespace duckdb
